+++ Example/AndesUI/AndesUI-demoapp-Bridging-Header.h
@@ -6,3 +6,4 @@
 #import "AndesUIViewController.h"
 #import "MessageObjCViewController.h"
 #import "TextFieldObjCViewController.h"
+#import "BadgeObjCViewController.h"