+++ Example/AndesUI/Home/Routers/HomeRouter.swift
@@ -12,6 +12,7 @@ protocol HomeRouter: NSObject {
     func route(from: UIViewController)
     func routeToButton()
     func routeToMessages()
+    func routeToBadges()
     func routeToWhatsNew()
     func routeTextField()
 }
@@ -22,6 +23,7 @@ class HomeAppRouter: NSObject {
 
     let buttonsRouter = ButtonsAppRouter()
     let messagesRouter = MessagesAppRouter()
+    let badgesRouter = BadgesAppRouter()
     let whatsNewRouter = WhatsNewAppRouter()
     let textFieldRouter = TextFieldsAppRouter()
 }
@@ -44,6 +46,10 @@ extension HomeAppRouter: HomeRouter {
 
     func routeToMessages() {
         messagesRouter.route(from: view)
+    }
+
+    func routeToBadges() {
+        badgesRouter.route(from: view)
     }
 
     func routeToWhatsNew() {