+++ LibraryComponents/Classes/AndesBadge/Border/AndesBadgeBorderProtocol.swift
@@ -1,0 +1,8 @@
+//
+//  AndesBadgeBorderProtocol.swift
+//  AndesUI
+//
+
+internal protocol AndesBadgeBorderProtocol {
+    var corners: CACornerMask { get }
+}