+++ LibraryComponents/Classes/AndesBadge/Border/AndesBadgeBorderRounded.swift
@@ -1,0 +1,8 @@
+//
+//  AndesBadgeBorderRounded.swift
+//  AndesUI
+//
+
+internal struct AndesBadgeBorderRounded: AndesBadgeBorderProtocol {
+    public var corners: CACornerMask = [.layerMinXMaxYCorner]
+}