+++ LibraryComponents/Classes/AndesBadge/Border/AndesBadgeBorderStandard.swift
@@ -1,0 +1,8 @@
+//
+//  AndesBadgeBorderStandard.swift
+//  AndesUI
+//
+
+internal struct AndesBadgeBorderStandard: AndesBadgeBorderProtocol {
+    public var corners: CACornerMask = [.layerMinXMinYCorner, .layerMinXMaxYCorner, .layerMaxXMinYCorner, .layerMaxXMaxYCorner]
+}