+++ LibraryComponents/Classes/AndesBadge/Config/AndesBadgeViewConfigFactory.swift
@@ -1,0 +1,28 @@
+//
+//  AndesBadgeViewConfigFactory.swift
+//  AndesUI
+//
+//  Created by Ignacio Guri on 20/03/2020.
+//
+
+import Foundation
+
+internal class AndesBadgeViewConfigFactory {
+    static func provideInternalConfig(from badge: AndesBadge) -> AndesBadgeViewConfig {
+        let typeIns = AndesBadgeTypeFactory.provide(badge.type)
+        let hierarchyIns = AndesBadgeHierarchyFactory.provide(badge.hierarchy, forType: typeIns)
+
+        var config = AndesBadgeViewConfig(backgroundColor: hierarchyIns.backgroundColor)
+
+        if badge.modifier == .pill, let pillText = badge.text {
+            config = AndesBadgeViewConfig(backgroundColor: hierarchyIns.backgroundColor, text: pillText, size: badge.size, textColor: hierarchyIns.textColor, border: badge.border)
+        }
+
+//        TODO: uncomment when Notification modifier is implemented
+//        if badge.modifier == .notification {
+//            config = AndesBadgeViewConfig(backgroundColor: hierarchyIns.backgroundColor, size: badge.size, icon: typeIns.icon!, iconBackgroundColor: hierarchyIns.backgroundColor)
+//        }
+
+        return config
+    }
+}