+++ LibraryComponents/Classes/AndesBadge/Hierarchy/AndesBadgeHierarchyProtocol.swift
@@ -1,0 +1,11 @@
+//
+//  AndesBadgeHierarchyProtocol.swift
+//  AndesUI
+//
+
+import Foundation
+
+internal protocol AndesBadgeHierarchyProtocol {
+    var backgroundColor: UIColor { get }
+    var textColor: UIColor { get }
+}