+++ LibraryComponents/Classes/AndesBadge/Size/AndesBadgeSizeLarge.swift
@@ -1,0 +1,18 @@
+//
+//  AndesBadgeSizeLarge.swift
+//  AndesUI
+//
+
+internal struct AndesBadgeSizeLarge: AndesBadgeSizeProtocol {
+    public var font: UIFont = AndesStyleSheetManager.styleSheet.semiboldSystemFontOfSize(size: 12)
+
+    public var lineHeight: CGFloat = 1
+
+    public var height: CGFloat = 24
+
+    public var borderRadius: CGFloat = 12
+
+    public var verticalPadding: CGFloat = 7
+
+    public var horizontalPadding: CGFloat = 8
+}