+++ LibraryComponents/Classes/AndesBadge/Size/AndesBadgeSizeProtocol.swift
@@ -1,0 +1,13 @@
+//
+//  AndesBadgeSizeProtocol.swift
+//  AndesUI
+//
+
+internal protocol AndesBadgeSizeProtocol {
+    var font: UIFont { get }
+    var lineHeight: CGFloat { get }
+    var height: CGFloat { get }
+    var borderRadius: CGFloat { get }
+    var verticalPadding: CGFloat { get }
+    var horizontalPadding: CGFloat { get }
+}