+++ LibraryComponents/Classes/AndesBadge/Type/AndesBadgeType.swift
@@ -1,0 +1,15 @@
+//
+//  AndesBadgeType.swift
+//  AndesUI
+//
+
+import Foundation
+
+/// Used to define the colors of an AndesBadge
+@objc public enum AndesBadgeType: Int {
+    case neutral
+    case highlight
+    case success
+    case warning
+    case error
+}