+++ LibraryComponents/Classes/AndesBadge/Type/AndesBadgeTypeHighlight.swift
@@ -1,0 +1,12 @@
+//
+//  AndesBadgeTypeHighlight.swift
+//  AndesUI
+//
+
+import Foundation
+
+class AndesBadgeTypeHighlight: AndesBadgeTypeProtocol {
+    var primaryColor: UIColor = AndesStyleSheetManager.styleSheet.accentColor
+    var secondaryColor: UIColor = AndesStyleSheetManager.styleSheet.accentColor100
+    var icon: String? = AndesIcons.feedbackInfo16
+}