+++ LibraryComponents/Classes/AndesBadge/Type/AndesBadgeTypeNeutral.swift
@@ -1,0 +1,12 @@
+//
+//  AndesBadgeTypeNeutral.swift
+//  AndesUI
+//
+
+import Foundation
+
+class AndesBadgeTypeNeutral: AndesBadgeTypeProtocol {
+    var primaryColor: UIColor = UIColor.Andes.gray450
+    var secondaryColor: UIColor = UIColor.Andes.gray070
+    var icon: String?
+}