+++ LibraryComponents/Classes/AndesBadge/Type/AndesBadgeTypeWarning.swift
@@ -1,0 +1,12 @@
+//
+//  AndesBadgeTypeWarning.swift
+//  AndesUI
+//
+
+import Foundation
+
+class AndesBadgeTypeWarning: AndesBadgeTypeProtocol {
+    var primaryColor: UIColor = UIColor.Andes.orange500
+    var secondaryColor: UIColor = UIColor.Andes.orange100
+    var icon: String? = AndesIcons.feedbackWarning16
+}